moved classes into their respected files (and created such files)
