+++ src/Barrier.cpp
@@ -1,0 +1,9 @@
+//
+// Created by meeve on 11.09.2021.
+//
+
+#include "Barrier.h"
+
+Barrier::Barrier(int inx, int iny, int indx, int indy) // constructor
+    : x(inx), y(iny), dx(indx), dy(indy), arr(nullptr) {}
+
