+++ src/Key.h
@@ -1,0 +1,15 @@
+//
+// Created by meeve on 11.09.2021.
+//
+
+#ifndef SEMESTER3_KEY_H
+#define SEMESTER3_KEY_H
+
+// Actually not key but Hostage
+class Key {
+public:
+    int x;
+    int y;
+};
+
+#endif //SEMESTER3_KEY_H
