+++ src/Rocket.h
@@ -1,0 +1,15 @@
+//
+// Created by meeve on 11.09.2021.
+//
+
+#ifndef SEMESTER3_ROCKET_H
+#define SEMESTER3_ROCKET_H
+
+class Rocket {
+public:
+    int x;
+    int y;
+    int damage;
+};
+
+#endif //SEMESTER3_ROCKET_H
